Refactor script function into its own file
